refactor(rs-sdk-ffi): hoist MnemonicResolverHandle + MnemonicResolverCoreSigner from platform-wallet-ffi
Both types are SDK-protocol material — they implement
`key_wallet::signer::Signer` over a Swift-supplied mnemonic-resolver
vtable, with no wallet state, no `AssetLockManager`, no
`WalletManager`. The historical placement in `platform-wallet-ffi`
was incidental: that crate happened to host the first consumer.
Now `VTableSigner` (identity / platform-address signer) and
`MnemonicResolverCoreSigner` (asset-lock-derived Core ECDSA signer)
sit together as the SDK's complete FFI signer surface.

What moved
==========

- `mnemonic_resolver_core_signer.rs` (the signer + tests) →
  `rs-sdk-ffi/src/mnemonic_resolver_core_signer.rs`
- `MnemonicResolverHandle`, `MnemonicResolverVTable`,
  `MnemonicResolveCallback`, `MNEMONIC_RESOLVER_BUFFER_CAPACITY`,
  `mnemonic_resolver_result` module → new
  `rs-sdk-ffi/src/mnemonic_resolver.rs`
- The `dash_sdk_mnemonic_resolver_create` /
  `_destroy` FFI entry points moved with the resolver, retaining
  identical C ABI shape (Swift bindings unchanged).

What stays
==========

- `IdentityKeyPersisterHandle`, `PersistKeyArgs`,
  `PersistKeyCallback`, `PERSIST_KEY_*` constants stay in
  `platform-wallet-ffi/src/derive_and_persist_callbacks.rs`. The
  Keychain-write semantics (`wallet_id` scoping, identity-index
  slot tagging) are wallet-domain. Module renamed in spirit
  (header now describes the persister-only role) without renaming
  the `.rs` filename — keeping git history intact for the persister
  bits.

Wiring
======

- `MnemonicResolverHandle.ctx` and `.vtable` fields promoted from
  `pub(crate)` to `pub` so the cross-crate dispatch sites in
  platform-wallet-ffi (`sign_with_mnemonic_resolver.rs`,
  `derive_identity_key_at_slot.rs`, etc.) can still reach the
  vtable. The struct is `#[repr(C)]` and its layout is part of
  the C ABI either way; field visibility was an internal-Rust
  nicety with no external invariant attached.
- 8 consumers in `platform-wallet-ffi` updated to import from
  `rs_sdk_ffi::{MnemonicResolverHandle, …}` instead of
  `crate::derive_and_persist_callbacks::{…}`.
- The signer file's `dashcore::secp256k1::*` import becomes
  `key_wallet::dashcore::secp256k1::*` since rs-sdk-ffi doesn't
  carry `dashcore` as a direct dep (key-wallet re-exports it).
- `parse_mnemonic_any_language` import shifts to the existing
  `signer_simple::parse_mnemonic_any_language` already living in
  rs-sdk-ffi (was duplicated identically in
  platform-wallet-ffi/identity_keys_from_mnemonic.rs — that copy
  stays for the rest of platform-wallet-ffi's own consumers).

Net diff
========

-660 lines from platform-wallet-ffi (one full file deletion +
210-line strip from derive_and_persist_callbacks.rs); +660 lines
in rs-sdk-ffi as the two new files. Pure relocation, no behavior
change. Every consumer compiles against the new import path; iOS
build succeeds with regenerated headers (8 mnemonic-resolver
declarations now live in `rs-sdk-ffi/rs-sdk-ffi.h`, the 21
references in `platform-wallet-ffi.h` are all consumer-side
function parameters that take `MnemonicResolverHandle*`).
