+++ packages/rs-platform-wallet-ffi/src/derive_and_persist_callbacks.rs
@@ -1,211 +1,59 @@
-//! C-ABI callback handles backing the
-//! [`crate::dash_sdk_derive_and_persist_identity_keys`] entry point.
+//! Identity-key persister C-ABI callback handle backing the
+//! [`crate::dash_sdk_derive_and_persist_identity_keys`] entry point
+//! and friends.
 //!
 //! The architectural intent is captured in `swift-sdk/CLAUDE.md`'s
 //! "no mnemonic round-tripping" rule: derivation pipelines must live
 //! on the Rust side, with Swift exposing only the two operations
 //! Rust cannot perform from its side — reading the mnemonic from
 //! the iOS Keychain and writing the derived key bytes back. This
-//! module exposes those two operations as paired vtable + handle
-//! types so a Rust-owned derivation loop can call into Swift for
-//! each one.
+//! module exposes the **write** half (identity-key persister); the
+//! **read** half (mnemonic resolver) lives in
+//! [`rs_sdk_ffi::mnemonic_resolver`](::rs_sdk_ffi::mnemonic_resolver)
+//! so non-wallet SDK callers can share it.
 //!
-//! # Why two separate handles
+//! # Why split from the resolver
 //!
 //! Conceptually they could ride on a single vtable, but the
 //! lifetimes are different:
 //!
 //! - The mnemonic resolver is single-use per call (one read of one
-//!   wallet's mnemonic per `dash_sdk_derive_and_persist_identity_keys`
-//!   invocation).
+//!   wallet's mnemonic per derivation invocation).
 //! - The persister is multi-use per call (`key_count` writes per
 //!   invocation).
 //!
 //! Splitting them also makes future extensions cleaner: a hardware-
 //! key-store impl might persist via a Secure Enclave wrapper that has
-//! nothing in common with how the mnemonic is fetched.
+//! nothing in common with how the mnemonic is fetched. Keeping the
+//! persister wallet-side (rather than promoting it to rs-sdk-ffi)
+//! reflects that the Keychain-write semantics are wallet-domain:
+//! `wallet_id` scoping, identity-index slot tagging, etc.
 //!
 //! # Synchronous on purpose
 //!
 //! Unlike the
 //! [`SignerVTable`](rs_sdk_ffi::SignerVTable) infrastructure (which
 //! uses a `oneshot` channel + `CompletionSlot` because biometric
-//! prompts can take seconds), Keychain reads and writes are
-//! sub-millisecond, so both callbacks here are fully synchronous —
-//! no completion-callback ceremony, no `tokio::time::timeout`.
+//! prompts can take seconds), Keychain writes are sub-millisecond,
+//! so the persister callback is fully synchronous — no
+//! completion-callback ceremony, no `tokio::time::timeout`.
 //! The Rust side blocks the calling thread for the duration of one
 //! Keychain hop. Callers that already invoke the derivation FFI on
 //! a background queue (the iOS pattern) get exactly what they want.
 //!
 //! # Lifetime / cleanup
 //!
-//! The Swift side calls `dash_sdk_mnemonic_resolver_create` /
-//! `dash_sdk_identity_key_persister_create`, which return opaque
-//! `*mut MnemonicResolverHandle` / `*mut IdentityKeyPersisterHandle`
-//! pointers. Pass them to
-//! `dash_sdk_derive_and_persist_identity_keys`. Then call the
-//! matching `_destroy` function — that fires the supplied destructor
-//! callback (typically the `Unmanaged.fromOpaque(...).release()`
-//! pattern Swift uses for its `KeychainSigner` ctx) so the Swift
-//! object can drop.
+//! The Swift side calls [`dash_sdk_identity_key_persister_create`],
+//! which returns an opaque `*mut IdentityKeyPersisterHandle` pointer.
+//! Pass it to
+//! [`crate::dash_sdk_derive_and_persist_identity_keys`]. Then call
+//! [`dash_sdk_identity_key_persister_destroy`] — that fires the
+//! supplied destructor callback (typically the
+//! `Unmanaged.fromOpaque(...).release()` pattern Swift uses for its
+//! `KeychainSigner` ctx) so the Swift object can drop.
 
 use std::ffi::c_void;
 use std::os::raw::c_char;
-
-// ---------------------------------------------------------------------------
-// Mnemonic resolver — Rust → Swift "fetch BIP-39 mnemonic for wallet_id"
-// ---------------------------------------------------------------------------
-
-/// Maximum mnemonic length, in bytes (excluding the trailing NUL),
-/// the resolver buffer can hold.
-///
-/// 1024 bytes covers every BIP-39 wordlist with margin to spare:
-/// the longest 24-word mnemonics in any supported language come in
-/// well under that — Korean (the longest of the supported lists)
-/// tops out near 700 bytes for a 24-word phrase. Picking a power-
-/// of-two cap keeps the inline Rust stack buffer (used in
-/// [`crate::dash_sdk_derive_and_persist_identity_keys`]) cheap to
-/// zero-on-drop.
-pub const MNEMONIC_RESOLVER_BUFFER_CAPACITY: usize = 1024;
-
-/// Resolver result codes returned by [`MnemonicResolveCallback`].
-///
-/// Mirrors the success/failure shape of `PlatformWalletFFIResult`
-/// at a finer-grained level so the Rust side can distinguish
-/// "Swift hit the buffer cap" from "Swift had no mnemonic stored
-/// for this wallet".
-pub mod mnemonic_resolver_result {
-    /// Mnemonic copied into the buffer; `out_len` was set.
-    pub const SUCCESS: i32 = 0;
-    /// The Swift side has no mnemonic stored for this `wallet_id`.
-    /// Surfaced to the Rust caller as
-    /// `PlatformWalletFFIResult::ErrorWalletOperation` with a
-    /// "mnemonic missing" detail.
-    pub const NOT_FOUND: i32 = 1;
-    /// Mnemonic exceeded [`MNEMONIC_RESOLVER_BUFFER_CAPACITY`].
-    /// Should not happen in practice — the buffer is sized for
-    /// every BIP-39 wordlist's 24-word phrase plus margin.
-    pub const BUFFER_TOO_SMALL: i32 = 2;
-    /// Anything else (Keychain access denied, decode error, etc.).
-    /// Surfaced as `ErrorWalletOperation` with a generic detail
-    /// the Swift side hopefully logged before returning.
-    pub const OTHER: i32 = 3;
-}
-
-/// Function pointer type for the mnemonic-resolve callback.
-///
-/// # Wire shape
-///
-/// - `ctx`: opaque Swift-side context (typically the result of
-///   `Unmanaged.passRetained(swiftSelf).toOpaque()`). Passed back
-///   verbatim by Rust on every invocation.
-/// - `wallet_id_bytes`: pointer to a 32-byte wallet id. Valid for
-///   the duration of the call only.
-/// - `out_mnemonic_utf8`: writable buffer for the NUL-terminated
-///   UTF-8 mnemonic. Capacity is
-///   [`MNEMONIC_RESOLVER_BUFFER_CAPACITY`] bytes (room for the
-///   trailing NUL).
-/// - `out_capacity`: equal to [`MNEMONIC_RESOLVER_BUFFER_CAPACITY`].
-///   Surfaced explicitly so the implementation can sanity-check
-///   without assuming the constant.
-/// - `out_len`: receives the byte count written to
-///   `out_mnemonic_utf8`, EXCLUDING the trailing NUL. Must be set
-///   on success.
-///
-/// # Safety
-///
-/// - `out_mnemonic_utf8` must be valid for `out_capacity` writable
-///   bytes for the duration of this call.
-/// - `wallet_id_bytes` must be valid for 32 readable bytes.
-/// - `out_len` must be valid for one `usize` write.
-/// - The implementation MUST return one of the
-///   [`mnemonic_resolver_result`] codes; any other value is treated
-///   as `OTHER`.
-pub type MnemonicResolveCallback = unsafe extern "C" fn(
-    ctx: *const c_void,
-    wallet_id_bytes: *const u8,
-    out_mnemonic_utf8: *mut c_char,
-    out_capacity: usize,
-    out_len: *mut usize,
-) -> i32;
-
-/// C-compatible vtable for a mnemonic resolver.
-#[repr(C)]
-pub struct MnemonicResolverVTable {
-    /// Synchronous "fetch mnemonic for `wallet_id`".
-    pub resolve: MnemonicResolveCallback,
-    /// Destructor for the `ctx` pointer. Invoked exactly once when
-    /// the matching `dash_sdk_mnemonic_resolver_destroy` is called.
-    pub destroy: unsafe extern "C" fn(ctx: *mut c_void),
-}
-
-/// Opaque Rust-side handle to a Swift-owned mnemonic resolver.
-///
-/// Constructed by [`dash_sdk_mnemonic_resolver_create`], destroyed
-/// by [`dash_sdk_mnemonic_resolver_destroy`]. Pass the pointer
-/// into [`crate::dash_sdk_derive_and_persist_identity_keys`].
-#[repr(C)]
-pub struct MnemonicResolverHandle {
-    pub(crate) ctx: *mut c_void,
-    pub(crate) vtable: *mut MnemonicResolverVTable,
-}
-
-// SAFETY: Swift side promises both pointers are thread-stable for
-// the lifetime of this handle. The handle is owned by Swift
-// across the FFI; Rust just dereferences it under the same lock
-// shape the existing signer FFI uses.
-unsafe impl Send for MnemonicResolverHandle {}
-unsafe impl Sync for MnemonicResolverHandle {}
-
-/// Build a new resolver handle that wraps a Swift-owned `ctx`
-/// pointer plus a pair of function pointers.
-///
-/// # Safety
-/// - `resolve_callback` must conform to [`MnemonicResolveCallback`]'s
-///   contract.
-/// - `destroy_callback` must safely free `ctx` exactly once.
-/// - `ctx` may be null (the destructor will be invoked with null
-///   on destroy regardless).
-#[no_mangle]
-pub unsafe extern "C" fn dash_sdk_mnemonic_resolver_create(
-    ctx: *mut c_void,
-    resolve_callback: MnemonicResolveCallback,
-    destroy_callback: unsafe extern "C" fn(ctx: *mut c_void),
-) -> *mut MnemonicResolverHandle {
-    // Allocate the vtable on the heap so it has a stable address
-    // for the lifetime of the handle — same pattern
-    // `dash_sdk_signer_create_with_ctx` follows.
-    let vtable = Box::into_raw(Box::new(MnemonicResolverVTable {
-        resolve: resolve_callback,
-        destroy: destroy_callback,
-    }));
-    Box::into_raw(Box::new(MnemonicResolverHandle { ctx, vtable }))
-}
-
-/// Destroy a previously-created resolver handle.
-///
-/// Calls the supplied destructor exactly once with the original
-/// `ctx`, then frees the heap-allocated vtable and the handle box.
-/// Safe to call with a null pointer (no-op).
-///
-/// # Safety
-/// `handle` must have been produced by
-/// [`dash_sdk_mnemonic_resolver_create`] and must not have been
-/// destroyed already.
-#[no_mangle]
-pub unsafe extern "C" fn dash_sdk_mnemonic_resolver_destroy(handle: *mut MnemonicResolverHandle) {
-    if handle.is_null() {
-        return;
-    }
-    let owned = unsafe { Box::from_raw(handle) };
-    unsafe {
-        if !owned.vtable.is_null() {
-            ((*owned.vtable).destroy)(owned.ctx);
-            // Reclaim the vtable box.
-            let _ = Box::from_raw(owned.vtable);
-        }
-    }
-}
 
 // ---------------------------------------------------------------------------
 // Identity-key persister — Rust → Swift "save this derived key to Keychain"
@@ -342,6 +190,9 @@ pub unsafe extern "C" fn dash_sdk_identity_key_persister_create(
     persist_callback: PersistKeyCallback,
     destroy_callback: unsafe extern "C" fn(ctx: *mut c_void),
 ) -> *mut IdentityKeyPersisterHandle {
+    // Allocate the vtable on the heap so it has a stable address
+    // for the lifetime of the handle — same pattern
+    // `dash_sdk_signer_create_with_ctx` follows.
     let vtable = Box::into_raw(Box::new(IdentityKeyPersisterVTable {
         persist_key: persist_callback,
         destroy: destroy_callback,
@@ -349,7 +200,11 @@ pub unsafe extern "C" fn dash_sdk_identity_key_persister_create(
     Box::into_raw(Box::new(IdentityKeyPersisterHandle { ctx, vtable }))
 }
 
-/// Destroy a previously-created persister handle. Safe on null.
+/// Destroy a previously-created persister handle.
+///
+/// Calls the supplied destructor exactly once with the original
+/// `ctx`, then frees the heap-allocated vtable and the handle box.
+/// Safe to call with a null pointer (no-op).
 ///
 /// # Safety
 /// `handle` must have been produced by
@@ -366,6 +221,7 @@ pub unsafe extern "C" fn dash_sdk_identity_key_persister_destroy(
     unsafe {
         if !owned.vtable.is_null() {
             ((*owned.vtable).destroy)(owned.ctx);
+            // Reclaim the vtable box.
             let _ = Box::from_raw(owned.vtable);
         }
     }
@@ -380,35 +236,8 @@ mod tests {
     use super::*;
     use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 
-    static RESOLVER_CALLED: AtomicBool = AtomicBool::new(false);
-    static RESOLVER_DESTROYED: AtomicBool = AtomicBool::new(false);
     static PERSISTER_CALLS: AtomicUsize = AtomicUsize::new(0);
     static PERSISTER_DESTROYED: AtomicBool = AtomicBool::new(false);
-
-    unsafe extern "C" fn fake_resolve(
-        _ctx: *const c_void,
-        _wallet_id_bytes: *const u8,
-        out_buf: *mut c_char,
-        out_capacity: usize,
-        out_len: *mut usize,
-    ) -> i32 {
-        RESOLVER_CALLED.store(true, Ordering::SeqCst);
-        let phrase = b"abandon abandon abandon abandon abandon abandon abandon abandon \
-                       abandon abandon abandon about";
-        if phrase.len() + 1 > out_capacity {
-            return mnemonic_resolver_result::BUFFER_TOO_SMALL;
-        }
-        unsafe {
-            std::ptr::copy_nonoverlapping(phrase.as_ptr() as *const c_char, out_buf, phrase.len());
-            *out_buf.add(phrase.len()) = 0;
-            *out_len = phrase.len();
-        }
-        mnemonic_resolver_result::SUCCESS
-    }
-
-    unsafe extern "C" fn fake_resolver_destroy(_ctx: *mut c_void) {
-        RESOLVER_DESTROYED.store(true, Ordering::SeqCst);
-    }
 
     unsafe extern "C" fn fake_persist(_ctx: *const c_void, _args: *const PersistKeyArgs) -> u8 {
         PERSISTER_CALLS.fetch_add(1, Ordering::SeqCst);
@@ -417,36 +246,6 @@ mod tests {
 
     unsafe extern "C" fn fake_persister_destroy(_ctx: *mut c_void) {
         PERSISTER_DESTROYED.store(true, Ordering::SeqCst);
-    }
-
-    #[test]
-    fn resolver_create_destroy_roundtrip() {
-        unsafe {
-            let h = dash_sdk_mnemonic_resolver_create(
-                std::ptr::null_mut(),
-                fake_resolve,
-                fake_resolver_destroy,
-            );
-            assert!(!h.is_null());
-            assert!(!(*h).vtable.is_null());
-
-            // Drive the resolver once to confirm the wired-up callback fires.
-            let mut buf = [0u8; MNEMONIC_RESOLVER_BUFFER_CAPACITY];
-            let mut len: usize = 0;
-            let rc = ((*(*h).vtable).resolve)(
-                (*h).ctx,
-                [0u8; 32].as_ptr(),
-                buf.as_mut_ptr() as *mut c_char,
-                buf.len(),
-                &mut len,
-            );
-            assert_eq!(rc, mnemonic_resolver_result::SUCCESS);
-            assert!(len > 0);
-
-            dash_sdk_mnemonic_resolver_destroy(h);
-        }
-        assert!(RESOLVER_CALLED.load(Ordering::SeqCst));
-        assert!(RESOLVER_DESTROYED.load(Ordering::SeqCst));
     }
 
     #[test]
@@ -487,10 +286,9 @@ mod tests {
     }
 
     #[test]
-    fn destroy_handles_null() {
+    fn destroy_persister_null_handle() {
         unsafe {
-            // Should be no-ops.
-            dash_sdk_mnemonic_resolver_destroy(std::ptr::null_mut());
+            // Should be a no-op.
             dash_sdk_identity_key_persister_destroy(std::ptr::null_mut());
         }
     }
