+++ packages/rs-platform-wallet-ffi/src/lib.rs
@@ -44,7 +44,6 @@ pub mod identity_update;
 pub mod identity_withdrawal;
 pub mod managed_identity;
 pub mod manager;
-pub mod mnemonic_resolver_core_signer;
 pub mod manager_diagnostics;
 pub mod memory_explorer;
 pub mod persistence;
@@ -101,7 +100,6 @@ pub use identity_update::*;
 pub use identity_withdrawal::*;
 pub use managed_identity::*;
 pub use manager::*;
-pub use mnemonic_resolver_core_signer::*;
 pub use manager_diagnostics::*;
 pub use memory_explorer::*;
 pub use persistence::*;
